Move main logging to separate file.
Add ability to send log messages from react app to log file.
